Audio is starting to make noises other then weee and currrr

+++ source/audio.h
@@ -3,8 +3,8 @@
 #include <stdint.h>
 
 void init_audio();
-void play_effect(int16_t* data, uint32_t freq, uint32_t sample_length);
-void play_music(int16_t* data, uint32_t freq, uint32_t sample_length);//loops until told to stop
+void play_effect(int8_t* data, uint32_t freq, uint32_t sample_length);
+void play_music(int8_t* data, uint32_t freq, uint32_t sample_length);//loops until told to stop
 void stop_music();
 
 void play_test();//test function, plays a test noise
